requirements.txt
asgiref==3.3.1
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
Django==3.1.5
django-cors-headers==3.6.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-spectacular==0.17.2
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
packaging==20.8
Pillow==8.1.0
psycopg2==2.8.6
PyJWT==2.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-decouple==3.4
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.16.12
selenium==3.141.0
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
